requirements.txt
# LangChain and LLM
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# Memory and embeddings (optional for summary memory)
langchain-core>=0.3.0

# Tools
duckduckgo-search>=6.0.0
requests>=2.31.0

# Chat UI
streamlit>=1.28.0

# Environment
python-dotenv>=1.0.0
